Adding connection fail alert and button enabling-disabling
First, made changes in "SerialConnection" function to fix showing connection success message not only after connection established but after every situation when "OPENbtn" is clicked if the connection is open. But added enabling/disabling of buttons according to the current state of connection to prevent this bug.
Then, added an alert for connection failure, too.
Then, according to the LED state, also added enabling/disabling of LED buttons.