RFID Door Lock Implementation with LEDs
DumpInfo file is an example file from MFRC522 library, and I used it for connection of RFID reader module and usage of library while coding.
If the card is valid, then the servo turns and green led is on, and if the card is invalid, then red led is on. Also valid and invalid card ID is written on serial.
close #6