Line Follower Calibration V1
This code doesn't work. Because my sensors don't behave consistent. In the same conditions, while nothing is changed, one of them can behave on the farthest side of the range. It can happen only once in a time or in a period of time, but not again.
So this causes whole calculated values to turn upside down, and never work properly out of that period.
So to be able to use this code, I need sensors with more consistent behavior, and at least now, I don't have any.