Configuration Logic Coding
"configData" array is a member of the "Main" form class. "preConfig" array is a member of the "Configuration" form class. When the "configuration" button at the "Main" class is clicked, the data in the "configData" array is transferred to the "preConfig" array.
In the loading of the "configuration" form, the configuration list is appeared according to data in the "preConfig" array.
According to the checked "R_G_B" radio button and checked location radio button, when the "Match" button is clicked,  new matching appears on the list at the left-hand side. You can use the match button repeatedly. When the "Save" button is clicked, the configuration that appears on the list is saved on the "configLocation.txt" file. (locate at the debug directory)
And some user interface design changes.