Changes for configuration feature
This file is recoded according to the following scenario:
After the "Communication with robot arm information message" returns, configuration data is sent to the sensor Arduino board. "Configuration data received" message or "Configuration data received" message is expected. According to the incoming message, system status information and information messages are displayed and button active-inactive settings are made.
Also, the changes for the configure button for enabling until the system starts and after the system stops, and disabling after the system starts and until the system stops were done.