Changes for configuration at sensors side
When I added new lines for configuration data communication in the "loop" I realize that the loop is again getting complex or messy. So I placed the codes related to the initialization phase in a function.
After getting the message "ready" from the robot arm, then the sensor Arduino board waits for configuration data. After this data is received, the message indicates "data is received" is sent. The data is put in an array, and the system state is changed to the "start". If the configuration data is not received, then the system state remains pending, and the message indicates "data is not received" is sent.
Also, for decoding configuration data a function is added and used at the sending of the motion data after color detection.