Raw Analog Value of the sensors of the vehicle with fixed threshold
First, intended to test the vehicle movement with a fixed threshold, but as recorded to a video, while the values of the sensors are very variable, once around 40, then 60, after that 70 for white ground, I couldn't figure out how to determine a fixed threshold value. Also the day that the video is recorded, in a at least three-hour interval, I couldn't see the white ground value around 30, as I got before so many times.