Some Logic Mistakes Correction
In the previous code, the "stop" and forward" statement was not ended until it goes out the situation that enters there. But the intention is it does the "stop and forward" only twice then stop. Then only "forward", "turn left", "turn right" or "turn" statements occur, it moves. So the code corrected to act like this.