Some Logic Correcitons
This code is written assuming that this logic only turns one "logic 1" value. But I observe that sometimes the logic can get more than one "logic 1" value. If these values are like 00110 or 01100 then it turns two functions, left and forward and right and forward. so if(digital value[2]==1) was changed to control only digital 2 is 1 and others 0.
The else at the end of the control blocks was deleted. Because for this to work, all control blocks need to be linked together with "else if". Instead of doing this, I added all 0 controls.
Also, to hope to avoid motor noice, forward pins of the motors that should be stopped at turning are set to LOW. But it didn't work.