RFID Cards ID Saving as Previous and Current
First, the RFID card reading part of the code put in a function, and this function is called from the place that this part was located.
In this function, read card ID is  record in the readCardID array. Then authorized card control is done like before. And at the end, read card ID is recorded to the preCardID array. And readCardID array elements are equalized to 0.
RFID _DEBUG added to debug blocks. Debug blocks have been reworked.
RFID cards' IDs were recorded in a text file.
close #34