Start Up the Vehicle with Hıgh Power
First, the name of the control variable that uses for indicates that entered in extraordinary situations, is changed. Then, a new control variable, controlStop, has been added to determine which movement function comes from., and the value of the variable is assigned to 1.
Then, at the end of every movement function, the value of controlStop variable is assigned, and at the beginning of some movement functions (excluding stopping() function), according to the value of the control stop variable, entering the code blocks that are for push the vehicle is determined. close #36