For Calibration of Line Following
To determine the sensors that see black are whether far from the center or near to the center or the combination of the two sensors or the center sensor,  "positionCalculation" function was added. Every sensor was given a value. The values given to the sensors are chosen to be like that any summation of the combination of them isn't the same.
The summation of the values of the black sensors is turned back.
According to the summation value, turning power values of "moveRight" and "moveLeft" functions are determined.