Reading Card ID From Serial via Bluetooth
With the help of "controlInTimer" parameter, if "#" line is detected, the next line is assigned to "cardID" variable. But before assigning, the content of "cardID" variable is transferred to "preCardID" variable.
Some changes have been made to the user interface operation.
In the "Clear" button click event, after clearing the "rich text box", the previous and last read card ID values were coded to be print.