Designing "Direction Settings" Window
There is a group box for radio buttons of locations. There is another group box for functions. In this group box, there are textboxes, a label to show the vehicle location, four buttons, respectively for calculation of the path, showing IDs of the locations, with labels under radio buttons, and assigning new card IDs to locations.
The user selects a location, and select the radio button of the vehicle position, then the "calculation" button turns to "enabled", and then click on the button, it calculates the path, and writes the direction to the text box.
The user selects a location, and select the radio button for the new card ID. Then "Assign ID" button turns to enabled. Then enter the card ID into four textboxes, each of them for one number group of the ID. And click the button.
Show ID button shows the ID of each location, on labels that under radiobuttons.