Designing Changes
In the group box below, two radio buttons were added. And below group box test was turned to "Functions". The first radio button for path calculation. When this button is selected, the "Calculate" button is turned to enable. The second radio button is for assigning card IDs. When the second radio button is selected, the "Assign Card ID" button is turned to enable.
Under the first radio button, the information of the position of the vehicle, the previous position of the vehicle, and the information where the front of the vehicle is facing have been added.
Also, a text box that has a graph of the place, which the vehicle is moving was added. On the graph, the location of the vehicle is colored red. The destination location is also colored green. And with an arrow, where the front of the vehicle is facing is indicated.
Also, a "Reload" button is added for getting new data of the vehicle after moving.
To force the user to click "Reload" button "Calculate" button and the first radio button that for calculation is enabled after calculation.
But it can be still chosen "Card ID Assigning" radio button.