Path Calculation Codes (Results shown on graph and writtwn onthe text box)
In addition to the comments on the code :
If the previous destination and the current destination are not on the same line in the graph of the locations, the "Move forward" movement is returned to determine the position of the vehicle.
Path calculation is done in two-stage, if the destination and the current location are not on the same line. There is a function that calculates the direction, if the destination and the current location are on the same line. In two stage calculation, in the first stage current location is given to the function as start point, but as destination,the point that the intersection of the row of the destination and the column of the location is given. In the second stage this time, calculated point is given as start point, and the desired destination is also given.
Combination of the two calculations is the result of the two stage calculation.